Changed the vector allocation to allocating with a fixed capacity of the parsed content length if it exists, or 0.
Moved the code that copies the body part of the buffer into the vector into the if statement, so it runs only if there is a content length. This should improve performance
as no copying is done unless there is a body to begin with. If there is no body, the vector will have length 0 and no additional cost is incurred.