Added a custom variant for the responses enum to add flexibility for determining custom codes, messages, etc...
As a result, HTTPResponses is too big to store on the stack. Wrapped
the OK type and Error type in a Box pointer to store them on the heap.

Removed the generic T for the http result and renamed it to HTTPResult.
Now, HTTPResult is Ok(Box<HTTPResponses>) or Err(Box<HTTPResponses>)

Provided an http_ok and http_err function which accepts and a
HTTPResponses variable and creates Box pointer to it on the heap wrapped in
an ok and err variant respectivley.